Refactor drawing and saving interfaces
Saving is generic given a type indicating format.
Refactor draw now we don't need to return the self type.

This setup doesn't allow drawing more than once to the same canvas, and
hence doesn't allow animation. We'll get back to this.

Refactoring of Java2D code gives much cleaner code, with a lot more
reuse.

Basic implementation of drawing to SVG DOM. Doesn't properly style
images, but it's good enough for now.
