--- shared/src/main/scala/doodle/backend/Canvas.scala
@@ -1,8 +1,0 @@
-package doodle
-package backend
-
-import doodle.core.{DrawingContext, Image}
-
-trait Canvas[A] {
-  def draw(interpreter: Configuration => Interpreter, image: Image): A
-}
