+++ shared/src/main/scala/doodle/backend/Draw.scala
@@ -1,0 +1,8 @@
+package doodle
+package backend
+
+import doodle.core.{DrawingContext, Image}
+
+trait Draw {
+  def draw(interpreter: Configuration => Interpreter, image: Image): Unit
+}
