+++ shared/src/main/scala/doodle/backend/Formats.scala
@@ -1,0 +1,10 @@
+package doodle
+package backend
+
+object Formats {
+  sealed trait Png
+  sealed trait Gif
+  sealed trait Svg
+  sealed trait Pdf
+  sealed trait Screen
+}
