+++ shared/src/main/scala/doodle/backend/Save.scala
@@ -1,0 +1,8 @@
+package doodle
+package backend
+
+import doodle.core.Image
+
+trait Save[Format] {
+  def save[F <: Format](fileName: String, interpreter: Configuration => Interpreter, image: Image): Unit
+}
