+++ shared/src/main/scala/doodle/backend/package.scala
@@ -7,4 +7,5 @@ package object backend {
   type Interpreter = Image => Renderable
   type Metrics = (Font, String) => BoundingBox
   type Configuration = (DrawingContext, Metrics)
+  type Drawable = (Configuration => Interpreter, Image)
 }
