+++ shared/src/main/scala/doodle/syntax/ImageSyntax.scala
@@ -1,11 +1,14 @@
 package doodle.syntax
 
 import doodle.core.Image
-import doodle.backend.{Canvas, Interpreter, Configuration}
+import doodle.backend.{Draw, Interpreter, Configuration, Save}
 
 trait ImageSyntax {
   implicit class ToImageOps(val image: Image) {
-    def draw[A](implicit canvas: Canvas[A], interpreter: Configuration => Interpreter): A =
-      canvas.draw(interpreter, image)
+    def draw(implicit draw: Draw, interpreter: Configuration => Interpreter): Unit =
+      draw.draw(interpreter, image)
+
+    def save[Format](fileName: String)(implicit save: Save[Format], interpreter: Configuration => Interpreter): Unit =
+      save.save[Format](fileName, interpreter, image)
   }
 }
